Create pages folder, separate pages from components
