--- src/components/Projects.jsx
@@ -1,7 +1,0 @@
-import ProjectsSlide from "./ProjectsSlide"
-
-export const Projects = () => {
-  return (
-    <ProjectsSlide/>
-  )
-}
