+++ src/pages/Objectives.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Footer } from '../components/Footer'
+import { VisionMission } from '../components/VisionMission'
+
+const Objectives = () => {
+  return (
+    <div>
+      <VisionMission />
+      <Footer />
+    </div>
+  )
+}
+
+export default Objectives
