+++ src/pages/Projects.jsx
@@ -1,0 +1,11 @@
+import { Footer } from "../components/Footer";
+import ProjectsSlide from "../components/ProjectsSlide";
+
+export const Projects = () => {
+  return (
+    <div>
+      <ProjectsSlide />
+      <Footer />
+    </div>
+  );
+};
